# Are JVM defaults good enough for our purposes?