… .azure-pipelines-templates/trace_validation.yml