We seem to be using a lot of flags to tell the JVM about memory, and I am not convinced they are needed for our usage at this moment in time.