@heidihoward 2023-09-29
These options are also set in https://github.com/microsoft/CCF/blob/main/.github/workflows/tlaplus.yml and https://github.com/microsoft/CCF/blob/main/tla/install_deps.py