@lemmy 2023-09-30
A few years ago, I ran performance measurements with `-XX:+UseParallelGC` and saw a noticeable performance benefit compared to the default, latency-optimized GC. Is it possible that newer JVM versions have eliminated this performance gap?