@achamayou 2023-10-02
> Is it possible that newer JVM versions have eliminated this performance gap?

I've tried looking through JRE releases notes, and they were surprisingly painful to gather. There's been a lot of changes, but at least the defaults for Xms and Xmx have been revised several times, and are platform dependent too.

Simulation time appears unchanged:

Before: https://github.com/microsoft/CCF/actions/runs/6272070572 - 10 min 35 s
After: https://github.com/microsoft/CCF/actions/runs/6356023869 - 10 min 28 s

Same with trace validation:

Before: https://dev.azure.com/MSRC-CCF/CCF/_build/results?buildId=76603&view=logs&j=0c81e209-1101-513f-1f15-37e9c3575eb2 - 19 s
After: https://dev.azure.com/MSRC-CCF/CCF/_build/results?buildId=76681&view=logs&j=0c81e209-1101-513f-1f15-37e9c3575eb2&t=a05664f3-ab91-5eb2-6acd-14e2179ad708 - 18 s

Model checking does seem to have gone up:

Before: https://github.com/microsoft/CCF/actions/runs/6273245621 - 22 min 13 s
After: https://github.com/microsoft/CCF/actions/runs/6356023869 - 31 min 49 s

I'll do a few more runs to see if this is conclusively related (we use a default public pool for that job, performance isn't particularly consistent), but there isn't a good reason for these flags at least for simulation and tracing.