@achamayou 2023-10-02
Looking at the cost of a few options:

`-coverage`

**With:** `time java -Dtlc2.TLC.ide=Github -Dutil.ExecutionStatisticsCollector.id=be29f6283abeed2fb1fd0be898bc6601 -cp tla2tools.jar tlc2.TLC -workers auto -lncheck final -checkpoint 60 -coverage 60 -tool MCccfraft.tla 2>&1`

```
real    1m46.869s
user    13m53.430s
sys     0m1.609s
```

**Without:** `time java -Dtlc2.TLC.ide=Github -Dutil.ExecutionStatisticsCollector.id=be29f6283abeed2fb1fd0be898bc6601 -cp tla2tools.jar tlc2.TLC -workers auto -lncheck final -checkpoint 60 -tool MCccfraft.tla 2>&1`

```
real    1m1.045s
user    7m51.213s
sys     0m1.875s
```

`-checkpoints`

**Without:**

```
real    1m0.140s
user    7m43.943s
sys     0m1.923s
```

Little difference there, but probably still not very useful in CI unless we archive them.