@heidihoward 2023-10-02
> `-checkpoints`
> 
> **Without:**
> 
> ```
> real    1m0.140s
> user    7m43.943s
> sys     0m1.923s
> ```
> 
> Little difference there, but probably still not very useful in CI unless we archive them.

 `-checkpoint 60` means checkpoint once every 60 mins so I wouldn't expect it to make much difference when a model runs for 8 mins 
https://github.com/tlaplus/tlaplus/blob/master/general/docs/current-tools.md#command-line-options