@heidihoward 2023-10-02
> `-checkpoints`
> 
> **Without:**
> 
> ```
> real    1m0.140s
> user    7m43.943s
> sys     0m1.923s
> ```

By the way, I believe that removing `checkpoint 60` means that checkpointing reverts to the default of checkpointing every 30 mins instead of disabling checkpointing. I don't think there's any way to fully disable checkpointing.