@heidihoward 2023-10-02
Sorry for the long stream of replies

> Looking at the cost of a few options:
> 
> `-coverage`
> 
> **With:** `time java -Dtlc2.TLC.ide=Github -Dutil.ExecutionStatisticsCollector.id=be29f6283abeed2fb1fd0be898bc6601 -cp tla2tools.jar tlc2.TLC -workers auto -lncheck final -checkpoint 60 -coverage 60 -tool MCccfraft.tla 2>&1`
> 
> ```
> real    1m46.869s
> user    13m53.430s
> sys     0m1.609s
> ```
> 
> **Without:** `time java -Dtlc2.TLC.ide=Github -Dutil.ExecutionStatisticsCollector.id=be29f6283abeed2fb1fd0be898bc6601 -cp tla2tools.jar tlc2.TLC -workers auto -lncheck final -checkpoint 60 -tool MCccfraft.tla 2>&1`
> 

This roughly matches my experience. I ran a few tests locally and the difference between no coverage and `coverage 60` is roughly a factor of two. `coverage 60` is still fairly useful (at least during the development of specs) as the coverage info is available at the end of checking (even if checking took less than 60 mins) and I often manually check for actions that have 0 coverage unexpectedly. See screenshot below.

![image](https://github.com/microsoft/CCF/assets/1835251/d28c23ae-58af-433c-b095-67850b1af299)