@achamayou 2023-10-02
@heidihoward it sounds to me like `-coverage` can be set in manual runs then, for example through `_JAVA_OPTIONS`, but I don't think we should have it in the CI.