@achamayou 2023-10-02
> > Simulation time appears unchanged:
> > Before: https://github.com/microsoft/CCF/actions/runs/6272070572 - 10 min 35 s After: https://github.com/microsoft/CCF/actions/runs/6356023869 - 10 min 28 s
> 
> The simulation has a time budget of 10 mins, so I would not expect to see much change there

Doh, of course. So we do seem to be generating somewhat fewer states (~9%) post-change, 678329 vs 744363.