@achamayou 2023-10-02
@lemmy reverting to the default GC makes the runtime grow by about 15% for model checking, so I'll put that back.

```
time java -XX:+UseParallelGC -Dtlc2.TLC.ide=Github -Dutil.ExecutionStatisticsCollector.id=be29f6283abeed2fb1fd0be898bc6601 -cp tla2tools.jar tlc2.TLC -workers auto -lncheck final -tool MCccfraft.tla 2>&1

real    0m59.279s
user    7m36.310s
sys     0m2.693s

real    0m56.746s
user    7m16.754s
sys     0m2.681s

real    0m59.181s
user    7m36.114s
sys     0m2.561s

time java -Dtlc2.TLC.ide=Github -Dutil.ExecutionStatisticsCollector.id=be29f6283abeed2fb1fd0be898bc6601 -cp tla2tools.jar tlc2.TLC -workers auto -lncheck final -tool MCccfraft.tla 2>&1

real    1m2.915s
user    8m0.579s
sys     0m2.198s

real    1m1.451s
user    7m47.127s
sys     0m2.190s

real    1m2.192s
user    7m58.164s
sys     0m2.128s
```