@achamayou 2023-10-02
> Thanks for confirming. :-)

Sorry, that was a little trigger happy, but the proliferation of flags was making me a little nervous. I'm glad it was possible consolidate in #5702.